+++ src/command_handlers/handle_sync_command.rs
@@ -3,71 +3,14 @@ use std::{
     env,
     io::{self, Write},
     process::Command,
-    str,
 };
 
 use reqwest::blocking::{Client, RequestBuilder};
 
-mod handle_branch_command;
-mod handle_stack_command;
-mod persistence;
-mod types;
+use crate::persistence::{self, FileData};
 
-pub fn handle_stack_command(options: &[String]) -> Result<(), &'static str> {
-    let stack_name_result = handle_stack_command::extract_stack_name(options);
-
-    if stack_name_result.is_none() {
-        match handle_stack_command::print_current_stack() {
-            Ok(_) => return Ok(()),
-            Err(e) => return Err(e),
-        }
-    }
-
-    let stack_name = stack_name_result.unwrap();
-
-    let create_option = match handle_stack_command::extract_create_option(options) {
-        Some(_) => true,
-        None => false,
-    };
-
-    if create_option {
-        match handle_stack_command::create_stack(stack_name) {
-            Ok(_) => {}
-            Err(e) => return Err(e),
-        }
-    }
-
-    match handle_stack_command::checkout_stack(stack_name) {
-        Ok(_) => Ok(()),
-        Err(e) => Err(e),
-    }
-}
-
-pub fn handle_branch_command(options: &[String]) -> Result<(), &'static str> {
-    let branch_name = match handle_branch_command::extract_branch_name(options) {
-        Some(name) => name,
-        None => return Err("Missing branch name"),
-    };
-    let create_option = match handle_branch_command::extract_create_option(options) {
-        Some(_) => true,
-        None => false,
-    };
-
-    if create_option {
-        match handle_branch_command::create_branch(&branch_name) {
-            Ok(_) => {}
-            Err(e) => return Err(e),
-        }
-    }
-
-    match handle_branch_command::checkout_branch(branch_name) {
-        Ok(_) => Ok(()),
-        Err(e) => Err(e),
-    }
-}
-
-pub fn handle_sync_command() -> Result<(), &'static str> {
-    let file_data = match persistence::read_from_file::<types::FileData>("save.json") {
+pub fn handler() -> Result<(), &'static str> {
+    let file_data = match persistence::read_from_file::<FileData>("save.json") {
         Ok(loaded_stacks) => loaded_stacks,
         Err(_) => return Err("Error when reading file"),
     };
@@ -266,7 +209,7 @@ fn extract_repo_owner_and_name() -> Option<(String, String)> {
         .expect("La commande git a échoué");
 
     // Convertit la sortie du processus en une chaîne de caractères
-    let output_str = str::from_utf8(&output.stdout).ok()?;
+    let output_str = std::str::from_utf8(&output.stdout).ok()?;
 
     // Sépare l'URL en parties en utilisant le séparateur "\t" (tabulation)
     let parts: Vec<&str> = output_str.split('\t').collect();
